+++ zscaler/zia/models/cloud_firewall_ips_rules.py
@@ -29,7 +29,6 @@ from zscaler.zia.models import cloud_firewall_source_groups as source_groups
 from zscaler.zia.models import cloud_firewall_nw_service_groups as nw_service_groups
 from zscaler.zia.models import cloud_firewall_nw_service as nw_service
 from zscaler.zia.models import common as common
-from zscaler.zia.models import common as common
 
 
 class FirewallIPSrules(ZscalerObject):
@@ -57,7 +56,6 @@ class FirewallIPSrules(ZscalerObject):
             self.state = config["state"] if "state" in config else None
             self.description = config["description"] if "description" in config else None
             self.last_modified_time = config["lastModifiedTime"] if "lastModifiedTime" in config else None
-            self.last_modified_by = config["lastModifiedBy"] if "lastModifiedBy" in config else None
 
             self.locations = ZscalerCollection.form_list(
                 config["locations"] if "locations" in config else [], location_management.LocationManagement
@@ -125,6 +123,16 @@ class FirewallIPSrules(ZscalerObject):
             self.enable_full_logging = config["enableFullLogging"] if "enableFullLogging" in config else None
             self.predefined = config["predefined"] if "predefined" in config else False
             self.default_rule = config["defaultRule"] if "defaultRule" in config else False
+
+            if "lastModifiedBy" in config:
+                if isinstance(config["lastModifiedBy"], common.CommonBlocks):
+                    self.last_modified_by = config["lastModifiedBy"]
+                elif config["lastModifiedBy"] is not None:
+                    self.last_modified_by = common.CommonBlocks(config["lastModifiedBy"])
+                else:
+                    self.last_modified_by = None
+            else:
+                self.last_modified_by = None
         else:
             self.access_control = None
             self.id = None